+++ src/main/java/chess/service/ChessBoardService.java
@@ -1,10 +1,8 @@
 package chess.service;
 
 import chess.entity.PieceEntity;
-import chess.exception.NoSuchGameException;
 import chess.model.GameResult;
 import chess.model.board.Board;
-import chess.model.board.BoardFactory;
 import chess.model.dao.GameDao;
 import chess.model.dao.PieceDao;
 import chess.model.piece.PieceFactory;
@@ -24,18 +22,12 @@ public class ChessBoardService {
         this.gameDao = gameDao;
     }
 
-    public Board toBoard(List<PieceEntity> rawBoard) {
+    public static Board toBoard(List<PieceEntity> rawBoard) {
         return new Board(rawBoard.stream()
                 .collect(Collectors.toMap(
                         piece -> Position.from(piece.getPosition()),
                         piece -> PieceFactory.create(piece.getName()))
                 ));
-    }
-
-    public String getTurn(Long gameId) {
-        String turn = gameDao.findTurnByGameId(gameId)
-                .orElseThrow(() -> new NoSuchGameException("잘못된 ID 값입니다."));
-        return turn;
     }
 
     public boolean isKingDead(Long gameId) {
@@ -46,15 +38,5 @@ public class ChessBoardService {
     public GameResult getResult(Long gameId) {
         Board board = toBoard(pieceDao.findAllPiecesByGameId(gameId));
         return GameResult.from(board);
-    }
-
-    public void exitGame(Long gameId) {
-        gameDao.updateTurnByGameId(gameId, "end");
-    }
-
-    public void restartGame(Long gameId) {
-        pieceDao.deleteByGameId(gameId);
-        pieceDao.savePieces(BoardFactory.create(), gameId);
-        gameDao.updateTurnByGameId(gameId, "white");
     }
 }
